Remove legacypart and legacysystem
